pyproject.toml
[tool.poetry]
name = "langchain-101"
version = "0.1.0"
description = ""
authors = ["Ashutosh Pathak <iamashutoshpathak@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
